+++ src/colorer/editor/Outliner.h
@@ -55,13 +55,13 @@ public:
    */
   size_t itemCount();
 
-  void startParsing(int lno);
-  void endParsing(int lno);
-  void clearLine(int lno, String* line);
-  void addRegion(int lno, String* line, int sx, int ex, const Region* region);
-  void enterScheme(int lno, String* line, int sx, int ex, const Region* region, const Scheme* scheme);
-  void leaveScheme(int lno, String* line, int sx, int ex, const Region* region, const Scheme* scheme);
-  void modifyEvent(int topLine);
+  void startParsing(size_t lno);
+  void endParsing(size_t lno);
+  void clearLine(size_t lno, String* line);
+  void addRegion(size_t lno, String* line, int sx, int ex, const Region* region);
+  void enterScheme(size_t lno, String* line, int sx, int ex, const Region* region, const Scheme* scheme);
+  void leaveScheme(size_t lno, String* line, int sx, int ex, const Region* region, const Scheme* scheme);
+  void modifyEvent(size_t topLine);
 
 protected:
   bool isOutlined(const Region* region);
@@ -71,7 +71,7 @@ protected:
   std::vector<OutlineItem*> outline;
   bool lineIsEmpty;
   int curLevel;
-  int modifiedLine;
+  size_t modifiedLine;
 };
 
 #endif
